requirements.txt
pycryptodome>=3.20.0
pytest>=7.4.0